food class

+++ diabeasy-back/tblBelongToRecipe.cs
@@ -14,8 +14,8 @@ namespace diabeasy_back
     
     public partial class tblBelongToRecipe
     {
-        public Nullable<int> carbohydrates { get; set; }
-        public Nullable<int> sugars { get; set; }
+        public Nullable<double> carbohydrates { get; set; }
+        public Nullable<double> sugars { get; set; }
         public Nullable<int> weightInGrams { get; set; }
         public int UnitOfMeasure_id { get; set; }
         public int Recipe_id { get; set; }
